move column sort arrows to icons file
